# maefa-store/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Maefa Store
  version: 1.0.0
  description: |
    API REST du serveur Maefa Store (Node.js + Express). Toutes les réponses sont en JSON.
    Ouvrir ce fichier dans https://editor.swagger.io pour une page interactive.

    Accès :
    - **gérante** : en-tête `x-admin-pin` (jeton signé obtenu par `POST /api/admin/login`) ;
    - **cliente connectée** : en-tête `Authorization: Bearer <jeton>` (obtenu par `/api/auth/verify`) ;
    - **livreur** : jeton secret dans l'adresse (`/api/driver/{token}`, `/api/tour/{token}`).
servers:
  - url: https://maefa-store.onrender.com
  - url: http://localhost:8787
components:
  securitySchemes:
    admin: { type: apiKey, in: header, name: x-admin-pin }
    session: { type: http, scheme: bearer }
...
paths:
...
  /api/stats:
    get:
      tags: [Système]
      summary: 'Statistiques de visites'
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
...
  /api/showcase:
    get:
      tags: [Statut WhatsApp]
      summary: 'Vitrine des nouveautés (/s)'
      responses:
        '200': { description: OK }
    post:
      tags: [Statut WhatsApp]
      summary: 'Mettre à jour la vitrine'
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
...
  /api/voice/{slug}:
    get:
      tags: [Voix]
      summary: "Note vocale d'un produit"
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
    put:
      tags: [Voix]
      summary: 'Enregistrer une note vocale'
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
    delete:
      tags: [Voix]
      summary: 'Supprimer une note vocale'
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
...
  /api/notify/status:
    post:
      tags: [Notifications]
      summary: 'Notifier un changement de statut'
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
  /api/notify/restock:
    post:
      tags: [Notifications]
      summary: "Prévenir les clientes d'un retour en stock"
      security: [{ admin: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Accès refusé }
...
  /api/products/{id}/comments:
    get:
      tags: [Avis]
      summary: 'Avis et questions publiés sur une pièce, avec la note moyenne'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Messages publiés (prénom, texte, note, cliente vérifiée, réponse de Maefa) }
    post:
      tags: [Avis]
      summary: 'Publier un avis ou une question (liens et numéros refusés) ; « Cliente vérifiée » si la pièce a été livrée'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [author, text]
              properties:
                author: { type: string, maxLength: 40 }
                text: { type: string, maxLength: 600 }
                rating: { type: integer, minimum: 1, maximum: 5, nullable: true }
      responses:
        '201': { description: Publié }
        '400': { description: Prénom ou message manquant, lien ou numéro de téléphone }
        '429': { description: Trop de messages }
  /api/admin/comments:
    get:
      tags: [Avis]
      summary: 'Tous les messages (gérante)'
      security: [{ admin: [] }]
      responses: { '200': { description: OK } }
  /api/admin/comments/{id}:
    patch:
      tags: [Avis]
      summary: 'Répondre, masquer ou republier'
      security: [{ admin: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    delete:
      tags: [Avis]
      summary: 'Supprimer un message'
      security: [{ admin: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }